add authorization with apikey in headers

+++ api/training/delete.php
@@ -7,20 +7,25 @@ header('Content-Type: application/json');
   
 include_once '../../databaseConnection.php';
 include_once '../objects/training.php';
+include_once '../authorization.php';
 
 $database = new Database();
 $db = $database->getConnection();
 
 $training = new training($db);
+
+$headers = apache_request_headers();
   
 $training->id = isset($_GET['id']) ? $_GET['id'] : die();
 
-if($training->delete()){
-        http_response_code(201);
-        echo json_encode(array("message" => "training was deleted."));
-    }else{
-        http_response_code(503);
-        echo json_encode(array("message" => "Unable to delete training."));
-    }
-
+if(authorize($headers['Authorization']))
+{
+    if($training->delete()){
+            http_response_code(201);
+            echo json_encode(array("message" => "training was deleted."));
+        }else{
+            http_response_code(503);
+            echo json_encode(array("message" => "Unable to delete training."));
+        }
+}
 ?>
